Move the loader family out of storage/ into own package, /loaders and add index.
