+++ src/commands/schemas.test.ts
@@ -10,13 +10,11 @@ import {
   TEST_LOCAL_SCHEMA,
   TEST_LOCAL_SCHEMA_REGISTRY,
 } from "../../tests/unit/testResources";
+import { CCloudResourceLoader, LocalResourceLoader, ResourceLoader } from "../loaders";
 import { ContainerTreeItem } from "../models/main";
 import { Schema } from "../models/schema";
 import { SchemaRegistry } from "../models/schemaRegistry";
 import { KafkaTopic } from "../models/topic";
-import { CCloudResourceLoader } from "../storage/ccloudResourceLoader";
-import { LocalResourceLoader } from "../storage/localResourceLoader";
-import { ResourceLoader } from "../storage/resourceLoader";
 import {
   CannotLoadSchemasError,
   diffLatestSchemasCommand,
