+++ src/loaders/localResourceLoader.ts
