+++ src/loaders/resourceLoader.test.ts
