+++ src/loaders/resourceLoader.ts
