+++ src/viewProviders/topics.ts
@@ -8,6 +8,8 @@ import {
   localKafkaConnected,
 } from "../emitters";
 import { ExtensionContextNotSetError } from "../errors";
+import { ResourceLoader } from "../loaders";
+import { TopicFetchError } from "../loaders/loaderUtils";
 import { Logger } from "../logging";
 import { Environment } from "../models/environment";
 import { KafkaCluster } from "../models/kafkaCluster";
@@ -15,8 +17,6 @@ import { ContainerTreeItem } from "../models/main";
 import { isCCloud, isLocal } from "../models/resource";
 import { Schema, SchemaTreeItem, generateSchemaSubjectGroups } from "../models/schema";
 import { KafkaTopic, KafkaTopicTreeItem } from "../models/topic";
-import { TopicFetchError } from "../storage/loaderUtils";
-import { ResourceLoader } from "../storage/resourceLoader";
 
 const logger = new Logger("viewProviders.topics");
